+++ src/Presenter/TransactionPresenter.php
@@ -91,6 +91,11 @@ class TransactionPresenter implements PresenterInterface
                 ))) {
                     try {
                         $payment = $this->merchantClient->payments()->getPayment($paymentDetail->getId());
+                    } catch (\Exception $e) {
+                        $payment = $paymentDetails;
+                    }
+
+                    try {
                         $refunds = $this->merchantClient->payments()->getRefunds($paymentDetail->getId());
                         $captures = $this->merchantClient->payments()->getCaptures($paymentDetail->getId());
                         $paymentSpecificOutput = $this->getPaymentSpecificOutput(
@@ -150,7 +155,7 @@ class TransactionPresenter implements PresenterInterface
                             ];
                         }
                     }
-                    $refundableAmount = !$paymentDetails->getStatusOutput()->getIsRefundable() ? 0 : Tools::getRoundedAmountFromCents($totalCaptured - $totalRefunded + $totalPendingRefund, $currencyCode);
+                    $refundableAmount = !$paymentDetails->getStatusOutput()->getIsRefundable() ? 0 : Tools::getRoundedAmountFromCents($totalCaptured - $totalRefunded - $totalPendingRefund, $currencyCode);
                     $apiErrors = $paymentDetails->getStatusOutput()->getErrors() ?: [];
                     $errors = [];
                     foreach ($apiErrors as $apiError) {